refactor torrents state into app module to prepare for tui
