+++ src/torrent.rs
@@ -2,67 +2,33 @@
 //! torrent client, including loading METAINFO and
 //! making requests to trackers.
 
-use std::{fs, net::Ipv4Addr};
+use std::net::Ipv4Addr;
 
 use anyhow::{Context, Error};
-use rand::{Rng, distr::Alphanumeric};
 use serde_bencode::value::Value;
 use sha1::{Digest, Sha1};
 use urlencoding::encode_binary;
 
 use metainfo::MetaInfo;
 
-use download::tracker::TrackerRequest;
-
 mod download;
 mod metainfo;
-
-pub struct Torrents {
-    torrents: Vec<Torrent>,
-    peer_id: String,
-}
 
 pub struct Torrent {
     metainfo: MetaInfo,
     info_hash: String,
 }
 
-impl Torrents {
-    pub fn new() -> Self {
-        let prefix = b"-RS0001-";
-        let mut peer_id_bytes = [0u8; 20];
-
-        peer_id_bytes[..8].copy_from_slice(prefix);
-
-        let rand_part: String = rand::rng()
-            .sample_iter(&Alphanumeric)
-            .take(12)
-            .map(char::from)
-            .collect();
-
-        peer_id_bytes[8..].copy_from_slice(rand_part.as_bytes());
-
-        let peer_id = encode_binary(&peer_id_bytes).into_owned();
-
-        Torrents {
-            torrents: vec![],
-            peer_id,
-        }
-    }
-
-    /// Adds a torrent to the client from a .torrent `file_path`
-    pub fn add_torrent(&mut self, file_path: &str) -> Result<(), Error> {
-        let bytes: Vec<u8> = fs::read(file_path).expect("{file_path} not found.");
+impl Torrent {
+    /// Adds a torrent to the client from bytes loaded from a .torrent file.
+    pub fn load(bytes: &[u8]) -> Result<Self, Error> {
         let metainfo = MetaInfo::from_bytes(&bytes)?;
-
         let info_hash = Self::info_hash(&bytes)?;
 
-        self.torrents.push(Torrent {
+        Ok(Self {
             metainfo,
             info_hash,
-        });
-
-        Ok(())
+        })
     }
 
     /// Calculates an `info_hash` from the info dictionary bytes found in
@@ -72,7 +38,7 @@ impl Torrents {
     ///     - bytes are not valid bencode,
     ///     - info key is missing from bencode,
     ///     - an error happens converting back to bytes
-    pub fn info_hash(bytes: &[u8]) -> Result<String, Error> {
+    fn info_hash(bytes: &[u8]) -> Result<String, Error> {
         let value: Value = serde_bencode::from_bytes(&bytes)
             .context("Failed to decode .torrent file as bencode")?;
 
@@ -93,39 +59,37 @@ impl Torrents {
         Ok(encode_binary(&hash).into_owned())
     }
 
-    /// Basic WIP function to make a request to a tracker.
-    pub async fn download_torrents(&self) {
-        for torrent in &self.torrents {
-            let result = download::download(
-                &torrent.metainfo,
-                &TrackerRequest::new(&torrent.info_hash, &self.peer_id),
-            )
-            .await
-            .unwrap();
+    pub async fn download(&self, peer_id: &str) -> Result<(), Error> {
+        let result = download::download(
+            &self.metainfo,
+            &download::tracker::TrackerRequest::new(&self.info_hash, &peer_id),
+        )
+        .await?;
 
-            if let Ok(Value::Dict(mut map)) = serde_bencode::from_bytes::<Value>(&result[..]) {
-                if let Some(peers) = map.remove(&b"peers".to_vec()) {
-                    match peers {
-                        Value::Bytes(compact_peers) => {
-                            println!("Compact peers: {} bytes", compact_peers.len());
+        if let Ok(Value::Dict(mut map)) = serde_bencode::from_bytes::<Value>(&result[..]) {
+            if let Some(peers) = map.remove(&b"peers".to_vec()) {
+                match peers {
+                    Value::Bytes(compact_peers) => {
+                        println!("Compact peers: {} bytes", compact_peers.len());
 
-                            for chunk in compact_peers.chunks_exact(6) {
-                                let ip = Ipv4Addr::new(chunk[0], chunk[1], chunk[2], chunk[3]);
-                                let port = u16::from_be_bytes([chunk[4], chunk[5]]);
-                                println!("Peer: {}:{}", ip, port);
-                            }
-                        }
-                        Value::List(peer_list) => {
-                            println!("Non-compact peer list with {} entries", peer_list.len());
-                        }
-                        _ => {
-                            println!("Unexpected format for peers field: {:?}", peers);
+                        for chunk in compact_peers.chunks_exact(6) {
+                            let ip = Ipv4Addr::new(chunk[0], chunk[1], chunk[2], chunk[3]);
+                            let port = u16::from_be_bytes([chunk[4], chunk[5]]);
+                            println!("Peer: {}:{}", ip, port);
                         }
                     }
-                } else {
-                    println!("No peers field in response.");
+                    Value::List(peer_list) => {
+                        println!("Non-compact peer list with {} entries", peer_list.len());
+                    }
+                    _ => {
+                        println!("Unexpected format for peers field: {:?}", peers);
+                    }
                 }
+            } else {
+                println!("No peers field in response.");
             }
         }
+
+        Ok(())
     }
 }
